Move preview status helpers and hooks
