+++ assets/blocks/course-outline/index.js
@@ -5,4 +5,4 @@ export { default as OutlineBlock } from './outline-block';
 export { default as ModuleBlock } from './module-block';
 export { default as LessonBlock } from './lesson-block';
 import './store';
-import './status-store';
+import './status-preview/status-store';
