+++ assets/blocks/course-outline/outline-block/outline-edit.js
@@ -2,55 +2,25 @@
  * WordPress dependencies
  */
 import { InnerBlocks } from '@wordpress/block-editor';
-import { useSelect, useDispatch } from '@wordpress/data';
 import { compose } from '@wordpress/compose';
+import { useDispatch, useSelect } from '@wordpress/data';
 import { createContext, useEffect } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import OutlinePlaceholder from './outline-placeholder';
-import { COURSE_STORE } from '../store';
-import { useBlocksCreator } from '../use-block-creator';
 import OutlineSettings from './outline-settings';
 import { withDefaultBlockStyle } from '../../../shared/blocks/settings';
-import { COURSE_STATUS_STORE } from '../status-store';
 import ToggleLegacyCourseMetaboxesWrapper from '../../toggle-legacy-course-metaboxes-wrapper';
+import { useCourseLessonsForPreviewStatus } from '../status-preview/use-course-lessons-for-preview-status';
+import { COURSE_STORE } from '../store';
+import { useBlocksCreator } from '../use-block-creator';
 
 /**
  * A React context which contains the attributes and the setAttributes callback of the Outline block.
  */
 export const OutlineAttributesContext = createContext();
-
-/**
- * A hook to update the status store when a lesson is removed.
- *
- * @param {string}  clientId  The outline block id.
- * @param {boolean} isPreview Whether the block is currently in preview mode.
- */
-const useSynchronizeLessonsOnUpdate = function ( clientId, isPreview ) {
-	const outlineDescendants = useSelect(
-		( select ) => {
-			return select( 'core/block-editor' ).getClientIdsOfDescendants( [
-				clientId,
-			] );
-		},
-		[ clientId ]
-	);
-
-	const { stopTrackingRemovedLessons } = useDispatch( COURSE_STATUS_STORE );
-
-	useEffect( () => {
-		if ( ! isPreview ) {
-			stopTrackingRemovedLessons( outlineDescendants );
-		}
-	}, [
-		clientId,
-		outlineDescendants,
-		isPreview,
-		stopTrackingRemovedLessons,
-	] );
-};
 
 /**
  * Edit course outline block component.
@@ -80,8 +50,7 @@ const OutlineEdit = ( props ) => {
 		[ clientId ]
 	);
 
-	useSynchronizeLessonsOnUpdate( clientId, attributes.isPreview );
-
+	useCourseLessonsForPreviewStatus( clientId, attributes.isPreview );
 	const content = isEmpty ? (
 		<OutlinePlaceholder
 			addBlock={ ( type ) => setBlocks( [ { type } ], true ) }
