+++ assets/blocks/course-outline/status-preview/index.js
@@ -1,0 +1,8 @@
+/**
+ * An enum of Status constants.
+ */
+export const Status = {
+	NOT_STARTED: 'not-started',
+	IN_PROGRESS: 'in-progress',
+	COMPLETED: 'completed',
+};
