+++ assets/blocks/course-outline/status-preview/status-control/index.test.js
@@ -3,11 +3,12 @@
  */
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import { Status } from '../index';
 
 /**
  * Internal dependencies
  */
-import { StatusControl, Status } from './index';
+import { StatusControl } from './index';
 
 describe( '<StatusControl />', () => {
 	it( 'Should display the default options', () => {
