+++ assets/blocks/course-outline/status-preview/status-store.js
@@ -2,12 +2,13 @@
  * WordPress dependencies
  */
 import { select as selectData, registerStore } from '@wordpress/data';
+import { createReducerFromActionMap } from '../../../shared/data/store-helpers';
 
 /**
  * Internal dependencies
  */
-import { createReducerFromActionMap } from '../../shared/data/store-helpers';
-import { Status } from './status-control';
+
+import { Status } from './index';
 import { select, controls } from '@wordpress/data-controls';
 
 const DEFAULT_STATE = {
@@ -20,7 +21,7 @@ const DEFAULT_STATE = {
  */
 const actions = {
 	/**
-	 * Sets the status of a lesson.
+	 * Sets thecreateReducerFromActionMap status of a lesson.
 	 *
 	 * @param {string} lessonId The lesson id.
 	 * @param {string} status   The lesson status.
