+++ assets/blocks/course-progress-block/course-progress-edit.js
@@ -17,7 +17,7 @@ import {
 	withColorSettings,
 	withDefaultColor,
 } from '../../shared/blocks/settings';
-import { COURSE_STATUS_STORE } from '../course-outline/status-store';
+import { COURSE_STATUS_STORE } from '../course-outline/status-preview/status-store';
 import CourseProgressSettings from './course-progress-settings';
 import ToggleLegacyCourseMetaboxesWrapper from '../toggle-legacy-course-metaboxes-wrapper';
